+++ cognite/neat/_constants.py
@@ -5,6 +5,32 @@ from rdflib import DCTERMS, OWL, RDF, RDFS, SKOS, XSD, Namespace, URIRef
 
 from cognite import neat
 
+
+def _is_in_notebook() -> bool:
+    try:
+        from IPython import get_ipython
+
+        if "IPKernelApp" not in get_ipython().config:  # pragma: no cover
+            return False
+    except ImportError:
+        return False
+    except AttributeError:
+        return False
+    return True
+
+
+def _is_in_browser() -> bool:
+    try:
+        from pyodide.ffi import IN_BROWSER  # type: ignore [import-not-found]
+    except ModuleNotFoundError:
+        return False
+    return IN_BROWSER
+
+
+IN_PYODIDE = _is_in_browser()
+IN_NOTEBOOK = _is_in_notebook() or IN_PYODIDE
+
+
 PACKAGE_DIRECTORY = Path(neat.__file__).parent
 COGNITE_MODELS = (
     DataModelId("cdf_cdm", "CogniteCore", "v1"),