+++ cognite/neat/_issues/_base.py
@@ -11,7 +11,12 @@ from typing import Any, ClassVar, Literal, TypeAlias, TypeVar, get_args, get_ori
 from warnings import WarningMessage
 
 import pandas as pd
-from cognite.client.data_classes.data_modeling import ContainerId, PropertyId, ViewId
+from cognite.client.data_classes.data_modeling import (
+    ContainerId,
+    DataModelId,
+    PropertyId,
+    ViewId,
+)
 from pydantic_core import ErrorDetails
 
 from cognite.neat._utils.spreadsheet import SpreadsheetRead
@@ -67,7 +72,7 @@ class NeatIssue:
     extra: ClassVar[str | None] = None
     fix: ClassVar[str | None] = None
 
-    def as_message(self) -> str:
+    def as_message(self, include_type: bool = True) -> str:
         """Return a human-readable message for the issue."""
         template = self.__doc__
         if not template:
@@ -79,8 +84,10 @@ class NeatIssue:
             msg += "\n" + self.extra.format(**variables)
         if self.fix:
             msg += f"\nFix: {self.fix.format(**variables)}"
-        name = type(self).__name__
-        return f"{name}: {msg}"
+        if include_type:
+            name = type(self).__name__
+            msg = f"{name}: {msg}"
+        return msg
 
     def _get_variables(self) -> tuple[dict[str, str], bool]:
         variables: dict[str, str] = {}
@@ -123,6 +130,8 @@ class NeatIssue:
             return value.dump()
         elif isinstance(value, PropertyId):
             return value.dump(camel_case=True)
+        elif isinstance(value, DataModelId):
+            return value.dump(camel_case=True, include_type=False)
         raise ValueError(f"Unsupported type: {type(value)}")
 
     @classmethod
@@ -169,6 +178,8 @@ class NeatIssue:
             return tuple(cls._load_value(subtype, item) for item in value)
         elif type_ is ViewId:
             return ViewId.load(value)
+        elif type_ is DataModelId:
+            return DataModelId.load(value)
         elif type_ is PropertyId:
             return PropertyId.load(value)
         elif type_ is ContainerId:
@@ -266,7 +277,7 @@ class DefaultPydanticError(NeatError, ValueError):
             msg=error["msg"],
         )
 
-    def as_message(self) -> str:
+    def as_message(self, include_type: bool = True) -> str:
         if self.loc and len(self.loc) == 1:
             return f"{self.loc[0]} sheet: {self.msg}"
         elif self.loc and len(self.loc) == 2:
@@ -307,7 +318,7 @@ class RowError(NeatError, ValueError):
             url=str(url) if (url := error.get("url")) else None,
         )
 
-    def as_message(self) -> str:
+    def as_message(self, include_type: bool = True) -> str:
         input_str = str(self.input) if self.input is not None else ""
         input_str = input_str[:50] + "..." if len(input_str) > 50 else input_str
         output = (
@@ -350,7 +361,7 @@ class DefaultWarning(NeatWarning):
             source=warning.source,
         )
 
-    def as_message(self) -> str:
+    def as_message(self, include_type: bool = True) -> str:
         return str(self.warning)
 
 
@@ -421,7 +432,12 @@ class IssueList(NeatIssueList[NeatIssue]):
     def _repr_html_(self) -> str | None:
         if not self:
             return "<p>'No issues found'</p>"
-        return super()._repr_html_()
+        df = self.to_pandas()
+        agg_df = df["NeatIssue"].value_counts().to_frame()
+        if len(agg_df) < 10:
+            return agg_df._repr_html_()  # type: ignore[operator]
+        else:
+            return agg_df.head()._repr_html_()  # type: ignore[operator]
 
 
 T_Cls = TypeVar("T_Cls")